requirements.txt
torch
sentence-transformers
pyyaml
python-box
python-dotenv